# homework3-1/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Planetarium API",
    "description": "See planets, get planet details",
    "contact": {
      "email": "adzajko@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0-oas3"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/accounts/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get an Account's info.",
        "description": "Get Account by ID.",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "description": "Bad input parameters!"
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Edit an account's info.",
        "description": "Edit an account's info.",
        "operationId": "editAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account was succesfully edited."
          },
          "400": {
            "description": "Missing input!"
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Delete an account.",
        "description": "Delete an account that has the ID.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account was deleted."
          },
          "400": {
            "description": "Invalid parameter for ID!"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlanetInAccount": {
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Mars"
          }
        }
      },
      "Account": {
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "username": {
            "type": "string",
            "example": "goce_ars"
          },
          "password": {
            "type": "string",
            "example": "goce123"
          },
          "planets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanetInAccount"
            }
          }
        }
      }
    }
  }
}